[187] bgtypeid should be a uint32

+++ src/server/game/Server/Protocol/Handlers/BattleGroundHandler.cpp
@@ -336,7 +336,7 @@ void WorldSession::HandleBattlefieldListOpcode(WorldPacket &recv_data)
 {
     sLog->outDebug("WORLD: Recvd CMSG_BATTLEFIELD_LIST Message");
 
-    uint8 bgTypeId;
+    uint32 bgTypeId;
     recv_data >> bgTypeId;                                  // id from DBC
 
     BattlemasterListEntry const* bl = sBattlemasterListStore.LookupEntry(bgTypeId);
